app/requirements.txt
requests==2.23.0
beautifulsoup4==4.9.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Flask-Mail==0.9.1
Flask-Bcrypt==0.7.1

psycopg2-binary

PyJWT==1.7.1
celery==4.4.2
redis
gunicorn==20.0.4

Flask-Cors==3.0.8

scikit-learn==0.23.1